requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.4
django-autoslug==1.9.8
djangorestframework==3.12.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
pika==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
websockets==9.1
zope.interface==5.4.0
